No3/requirement.txt
kafka-python
pyspark~=2.4.6